backend/requirements.txt
langchain
wikipedia
langchain-community
langchain-openai
langchain-anthropic
python-dotenv
pydantic
duckduckgo-search
ddgs
fastapi
uvicorn
beautifulsoup4
requests
